+++ rumqttc/examples/tls_native.rs
@@ -2,28 +2,25 @@
 use std::error::Error;
 
 use rumqttc::{AsyncClient, Event, Incoming, MqttOptions, Transport};
-use tokio_rustls::rustls::ClientConfig;
+use tokio_native_tls::native_tls;
 
 #[tokio::main(flavor = "current_thread")]
 async fn main() -> Result<(), Box<dyn Error>> {
     pretty_env_logger::init();
     color_backtrace::install();
 
-    let mut mqttoptions = MqttOptions::new("test-1", "mqtt.example.server", 8883);
+    let mut mqttoptions = MqttOptions::new("test-1", "test.mosquitto.org", 8883);
     mqttoptions.set_keep_alive(std::time::Duration::from_secs(5));
-    mqttoptions.set_credentials("username", "password");
 
-    // Use rustls-native-certs to load root certificates from the operating system.
-    let mut root_cert_store = tokio_rustls::rustls::RootCertStore::empty();
-    root_cert_store.add_parsable_certificates(
-        rustls_native_certs::load_native_certs().expect("could not load platform certs"),
-    );
+    // Use native-tls to load root certificates from the operating system.
+    let mut builder = native_tls::TlsConnector::builder();
+    let pem = vec![1, 2, 3];
+    // let pem = include_bytes!("native-tls-cert.pem");
+    let cert = native_tls::Certificate::from_pem(&pem)?;
+    builder.add_root_certificate(cert);
+    let connector = builder.build()?;
 
-    let client_config = ClientConfig::builder()
-        .with_root_certificates(root_cert_store)
-        .with_no_client_auth();
-
-    mqttoptions.set_transport(Transport::tls_with_config(client_config.into()));
+    mqttoptions.set_transport(Transport::tls_with_config(connector.into()));
 
     let (_client, mut eventloop) = AsyncClient::new(mqttoptions, 10);